requirements.txt
torch
torchvision
Pillow
pandas
scikit-learn
tqdm

# Web Interface
gradio